--- tests/maxo_dialog/widgets/media/test_scroll2.py
@@ -1,24 +1,0 @@
-from unittest.mock import Mock
-
-from maxo.dialogs import DialogManager
-from maxo.dialogs.widgets.media import MediaScroll, StaticMedia
-from maxo.dialogs.widgets.text import Format
-
-
-async def test_render_media_scroll(mock_manager: DialogManager) -> None:
-    media = MediaScroll(
-        items=["0.png", "1.png", "2.png"],
-        media=StaticMedia(path=Format("/{item}")),
-        id="m",
-    )
-    res = await media.render_media(data={}, manager=mock_manager)
-    assert res.path == "/0.png"
-
-    await media.set_page(event=Mock(), page=2, manager=mock_manager)
-    res = await media.render_media(data={}, manager=mock_manager)
-    assert res.path == "/2.png"
-
-    managed = media.managed(mock_manager)
-    await managed.set_page(1)
-    res = await media.render_media(data={}, manager=mock_manager)
-    assert res.path == "/1.png"
